+++ TaskManagement/tachecomposite.h
@@ -10,7 +10,7 @@ class TacheComposite : public Tache
         TacheComposite(QString nom = "nouvelle tache", int duree = 0, double completion = 0.0, QList<Tache> suivantes = QList<Tache>(), QList<Tache> precedentes = QList<Tache>());
 
         int getDuree();
-        int getNum();
+        double getNum();
         double getCompletion();
 
         void setDuree(int duree);
